user_data: introduce implied_user_data()
Eventually instead of just tainting data as user_data, I want to say how
much the user can actually control.  For example, say we read a u8 from
the user but we pass it to a function and we also pass an unknown u32 to
the function.  In the current code, smatch assumes that the user can
control u32 bits of data but actually it is only a u8 of data that the
user controls.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>