extra: more limits on which variables are equivalent
If the assignment is:

	foo = &bar;

Then we don't record that as equivalent.  Unfortunately, I can't remember
why this is.  Probably it is because we don't record the state of &bar in
smatch_extra so they become out of sync.

If the assignment is:

	foo = foo->bar;

Then the problem is that after the assignment then the right side of the
assignment has changed so foo != foo->bar at the end.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>