struct_assignment: handle memcpy(foo, ...) where foo is not a struct
The code in question looks like this:

	__be32 foo = 0;
	memcpy(&foo, skb->data, sizeof(foo));

We want to say that "foo" is no longer zero so we create a fake assignment
for that.  Most of the time we won't know what the end value of "foo" is
but it's important to record that it changed.

I know that the name of this function said "struct" but this seemed like
the most natural place to put this.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>