type_val: store fake assignments in a slightly different format
This is a bit of a pointless thing, but whatever.  If it's a fake
assignment then I'm recording that the state is set to "min-max" instead
of "s32min-s32max".  That way when I do `smdb where struct_type member`
then it is easier to see the real assignments.

This hack has the other advantage that I don't need to modify any other
parts of the code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>