sparse: make bits_to_bytes round up instead of down
Currently, sparse handles arrays of bools incorrectly. If you declare an
array of bools like this:

    static _Bool boolarray[3] = {
        [0] = 1,
        [1] = 1,
    };

...you get warnings like this (which are bogus):

    ./test.c:2:10: warning: Initializer entry defined twice
    ./test.c:3:10:   also defined here

The problem is that bits_to_bytes rounds down instead of up, and sparse
defaults to _Bool being only 1 bit in size. This causes sparse to think
that they sit within the same byte, when they do not.

Fix bits_to_bytes to round up instead of down, and fix the call in
init_ctype to no longer correct for it. Also add a validation
test to ensure that we don't break this in the future.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>