comparison: introduce possible_comparison()
You call this function like this:

	if (possible_comparison(a, SPECIAL_EQUAL, b)) {

It returns true if we think that a is possibly equal to be.  I will use
this to check for off-by-one bugs like this:

	buf = kmalloc(len, GFP_KERNEL);
	...
	buf[len] = '\0';

If we store the relationship that "the size of buf is len" then later we
can see that the index is possibly equal so it's off by one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>