db: add a new table which hold parameter types
The immediate purpose of this table is that I want a way to look up which
parameters are void pointers.  If I were doing this from .c code then I
could look at the function definition but I want to do it after the files
have been parsed.

Also I think that this could just be useful information.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>