extra: handle merging NULL pointers better
Say you have:

	if (foo) {
		foo->bar = 1;
	} else {
		XXX
	}

In the original code when we came to the end of the else statement it would
have foo->bar as an unmatched state and it would set it to min-max.  Really
we should set it to the empty state.  That way at the end of the statement
we have foo->bar = 1 instead of foo->bar is unknown.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>