links: create a smatch_links.c file with functions for reverse links
We have two places now which use the concept of reverse links.  Say you
have code like:

	if (foo == bar) {

And you want to store in smatch_stored_conditions.c that the condition is
true or false, then you can do that.  But if either "foo" or "bar" changes
then you have to find the "foo == bar" state and set it to undefined.

So we have a state "foo == bar" and a link from "foo" to the state and link
from "bar" to the state.

This smatch_links.c file puts together some helper functions to make links
like this easier to use.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>