db: prefer literals to parameter math
Imagine you have code like validation/sm_inline1.c which does:

int frob(int *x)
{
        *x = *x * 3;
        return 0;
}

It was saying that if we call "*x = 1; frob(x);" then we set *x to
"3[$0 * 3]".  The caller first sets *x to 3 and then multiplies by
3 to get 9.

This works around that bug by not storing the parameter math.  If we're
not dealing with a literal then the parameter math fails because it doesn't
deal with range lists anyway.

This fix is the correct thing to do because why store duplicate data, and
it's how this is handled in smatch_db.c already.  However, we will hit this
bug again when smatch_parse_call_math.c is updated to handle ranges so we
will need an additional fix as well.  That fix will probably be trickier.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>