function_hooks: fix how database comparisons are done
The original code wasn't braiding together the strees correctly.  I have
basically copied how it's done for assignments.

The core problem which I was trying to address is that imagine you have
a function with three returns.

int cmp_x(int x, int y)
{
	if (x < y)
		return -1;
	if (x == y)
		return 0;
	return 1;
}

The caller does something like:

	if (cmp_x(x, 4) < 0) {

Then that means you have one true return, -1, and two false returns, 0 and
1.  We needed to reset the cur_stree back to the original state after we
had processed the first false return but we weren't doing that.  Instead we
were leaving x == 4, and when we asked which states were possible for a
1 return then it said that it's impossible.  If we reset the states back to
the original the it would say 5-s32max are possible.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>