ranges: use the ranges stored in the database properly(ish)
We are storing ranges in the database like:

	"s32min-s32max[<$0]"

We should be handling the comparison in the exact same way that
smatch_extra.c does.

This code has been tested lightly...

The only issue that I can think of right of the bat is that we aren't
actually type promoting the comparisons correct.  And, in fact, there is
not enough data in the database to be able to do that.  Oh well...  This
is an improvement.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>