parse_call_math: handle math like "$0->foo + 5"
Say you have a function that uses a struct member of a parameter to do
math instead of just a parameter.  The code could look like:

void frob(struct foo *p, int x)
{
	p->one = p->two + x;
}

This patch changes "p->two" into "$0->two" so it can be put into the
database and also changed back to "p->two" when we read from the database.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>