free: don't turn off warnings if --spammy is set
This change could mean that there are duplicate warnings when the --spammy
flag is used but that's ok.  In the original code, we could miss certain
types of bugs when the --spammy flag was used and that's not ok.

Also past experience says that it's a bad idea to turn off any warnings
when --spammy is used because I always run with --spammy on and might not
see false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>