flow: handle backwards goto hooks
I have had an idea for a long time that I need to do two passes and there
is already a --two-passes option.  But getting that to work seems like a
complicated step and this is an easy step small step because it's only
enabled for some functions and only when --info is turned on.

This is those functions where you have stuff like:

	if (ret)
		goto error;
out:
	return ret;
error:
	printk("i had an error\n");
	goto out;

People write this kind of ugly crap all the time and I don't know why.
I guess it must be a speed up on benchmarks?  Anyway, now we can parse
their code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>