db,extra: fixes to param filtering
I feel like maybe this code sucks and that maybe I just kept writing code
until my test case passed and then stopped without really understanding
the problem completely.

In the original code, you would cast the range lists but if you casted a
64 bit value that was larger than 2*32 to 32 bits then it ends up saying
that it could be any u32 value.  So that wasn't right.  Also I was using
the wrong type part of the time...

Anyway, I was trying to get kmalloc() to work like I wanted and it does so
I'm pushing this garbage.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>