db: build a new db first then overwrite the old one
In the old code, it deleted the old db and rebuilt it in place.  Now we
build smatch_db.sqlite.new and then move it to smatch_db.sqlite after it's
built.

The old way was annoying because you couldn't use the database for the hour
and a half while the database was being built.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>