delete the check_redundant_null_check.c
My thinking on this has evolved and I now feel that NULL checks can make
the code more readable.

Generally when you see a:

	if (foo)
		kfree(foo);

That is a sign of bad code.  Generally it is cause by "One Err" style error
handling where you do.

	foo = alloc();
	if (!foo)
		goto out;
	bar = alloc();
	if (!bar)
		goto out;

out:
	if (foo)
		kfree(foo);

This kind of error handling tends to have bugs, but the fix isn't to
remove the if statements, it's to unwind properly so that we know at every
line if foo is valid or not.

Also, I wrote this code a long time ago and at the time I thought it was
a clever hack but really I want to focus on bugs instead of style issues.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>