flow: fix how CALL_HOOK_AFTER is called
If the a function call is assigned then the call is processed with the
assignment.  In that case we need to delay calling FUNCTION_CALL_HOOK_AFTER
and call it with the assignment as well.

I moved the is_assigned_call() to smatch_flow.c so that everyone could use
it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>