core: introduce &ghost state
If "foo" is NULL then "foo->bar" is a ghost state.  When we merge a ghost
state with a &whatever then the result is &whatever.

Smatch_extra already has custom unmatched_state and merge functions so it
works this way.  But really this should be the default for everything.  I
didn't have a specific reason for this code, but it doesn't seem to cause
any false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>