function_hooks,db: calculate the return range based on the correct slist
We should look at the return_id and pop the fake stree before calculating
the return range.  Otherwise we end up calculating it based on the wrong
stree.

This is the correct thing to do and a bugfix but there is a second bug
here that is more tricky.  If we have:

	p = frob(p, xxx);

Where the return value is maybe "$0 + 4" then we end up calculating a new
return value for every entry in the database instead of having the same
return value every time.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>