function_hooks: improve how function assignments are handled
The problem here is this:

	p = frob(p, xxx);

The problem is that "p" is both the return range and a parameter.  When we
set "p" to the return values it messes up the code which needs to see the
parameter values.  So we store the ret_range in a separate place and only
set the state at the end.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>