impossible: introduce is_impossible_path() to say if a path is reasonable
A common source of false positives is code like this:

	foo = NULL;
	for (i = 0; i < limit; i++) {
		if (i != 0)
			foo = foo->next;
		else
			foo = bar;
	}

Smatch sees the "foo->next" dereference as a NULL deref.  The real way to
fix this is to do the multi-pass thing.  But another easier way to fix this
is to simply not print the error message if we are in an impossible state.

Each check needs to add this check individually.  For a lot of warnings we
still want to print the warning even if the state we are in is illogical.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>