fixup_kernel.sh: allow larger values in rw_verify_area()
I chose these values so that pos + count wouldn't overflow, so I just
randomly chose some largish values.  It turns out those values weren't
large enough.  The netxen_sysfs_validate_crb() function insists that the
offset be within a certain range that is outside the 0-1000000 range and
Smatch gets confused that it's not possible to get a successful return.

Anyway, 1000000000 + 1000000000 won't overflow the ints so it's ok to use
that here.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>