helper: change "&*foo" to "foo" at the lowest levels
I changed expr_to_str() to automatically simplify "&*foo" to "foo".  I
considered doing this in strip_expr() but for some reason when I looked at
that then it seemed like I would have to allocate a fake expression and I
didn't want to do that...  Doing the simplification here is sort of a new
approach for Smatch.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>