debug: add --trace=variable_name option
If you turn on --debug then it's too much output to understand.  This
prints how the state of a variable changes after each statement.  It
doesn't print all the merges so it's a bit limited.  And it looks at
the previous statement in terms of which one was analyzed last so if you
have something like:

	if (foo) {
		point a;
	} else {
		point b;
	}

Then after "point a" it will tell you that "foo" has changed to non-zero.
After "point b" it will tell you that "foo" has changed from non-zero to
zero.  So in that case the line numbers are off by one.  Oh well...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>