db: split success/fail returns as best we can
We try to split returns as much as possible but sometimes there are too
many states.  This is a compromise solution which says that in the kernel
if there are too many states to split every return, then at least lets
separate the negative error returns from the zero returns.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>