hooks: introduce a pre_merge hook (similar to a merge hook)
The problem this is trying to solve is, we want to look up something in
SMATCH_EXTRA to do the merge so we need the slist.  But the normal merge
function you don't/can't have the slist because you're dealing with two
different sm_states.

This is like the unmatched_state hook, but it's not unmatched, we actually
do have a state there.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>