common_functions: tweak strlen() so it doesn't go all the way up to s64max
Theoretically strlen() *can* go up to s64max, but in real life it's
difficult to arrange for that to happen in the kernel and it causes some
false positives.  Paper over that by saying the highest it can return is
101010.

Also the original code was using llong types and that's not correct.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>