db, extra: use PARAM_SET if the parameter is totally filtered out
If a function sets "*param = 42;" on some but not all paths, perhaps
something like this:


	if (*param > 42)
		*param = 42;

The we say that some of the original values are filtered and value "42" is
added.

But if we set *param on all paths then the filtered value is '' or the
empty set.  In this case I have added a new type called PARAM_SET.

The advantages of merging FILTER_VALUE and ADDED_VALUE into PARAM_SET are
that it makes the database smaller and simpler.  I had thought that if
fixes a bug, and actually it sort of does improve things it doesn't totally
fix the bug.  More improvements coming.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>