smatch: Add --enable option
It can be useful, especially when developing a new check, to only run
a subset of checks. The --enable option takes a comma-separated list
of integers or ranges and only enables those checks (the default is of
course that everything is enabled).

For now, anything called register_* is unconditionally called, since lots
of stuff seems to depend on that.