check_kernel_printf.c: Allow const char[] format arguments
If the format argument is a const char[] with a literal initializer we
might as well do the checking also in that case (gcc also does
that). Similarly if the format argument is const char* const (gcc
apparently does not do that). This removes around 100 messages from an
allmodconfig, and of course provides coverage of those same call
sites.

In most cases, however, I think it would be more appropriate to insert
the format string directly where it is used. The linker merges
identical string constants, so there's nothing saved by using an array
variable (on the contrary, this wastes a little space on the variable
name itself), and it is easier for a human to immediately verify that
everything is ok, instead of having to search back through the file to
see how foo is defined.