param_cleared, rosenberg: add underscored memset/cpy/zero() functions
On my kernel config we are only using the underscored versions of these
functions.  These weren't handled before so it leads to a lot of false
positive "information leak" warnings.

It's not clear to me why some of these are in a separate data file and
some are hard coded in the .c file.  Oh well...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>