struct_assignment: ignore more fake assignments
If we assign a struct to a struct: "a = b;" then it means that
"a->foo = b->foo;" etc.  There can be dozens of struct members like
"a->foo->bar->baz" and we record all these assignments multiple times.  It
takes a lot of data.

Most of the time we don't even care about any of the data so let's just
ignore it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>