buf_size: don't save useless information in the database
We're recording that we don't know the size of the buffer.  That's the
default state so we don't need to record it.  We're also recording that
the buf size is zero, that means the pointer is NULL.  This is saved in
smatch_extra so we don't need to duplicate it here.  There are a lot of
NULL pointers because stuff gets initialized with kzalloc() so this is a
lot of data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>