function_hooks: don't take a shortcut for "needless" states
The logic here was that if we don't have a state then we don't need to
set the state to unknown, since the default state is already unknown.  This
was supposed to save memory.

But these days there are a lot of places which tie into modification hooks
to tell when a variable is initialized so this information is still useful.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>