extra: separate PARAM_LIMIT and PARAM_FILTER
PARAM_LIMIT means we don't call extra_mod_hooks.
PARAM_FILTER means that we do.

I used to think that it was the right thing to call extra_mod_hooks when
PARAM_ADD is called.  This happens for every FILTER.  But it messes up
smatch_type_value.c because say we have code like:

	foo(&foo->bar, xxx);

And we filter some values and add some values, then both of those need to
be recorded and not just the added values.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>