strings: record which strings are being used
This is a lazy partial way of recording which strings are being used.  If
you have "strcpy(foo, "whatever");" then it stores state "whatever" in
"foo".

You can get the list of possible strings using the get_strings() function.
It is returned as a slist.  You need to check the slist to see if it has
the &undefined state.  It's could also be &merged, of course but I imagine
that searching for &undefined is normally enough.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>