kernel_printf: silence some false positives using smatch_strings.c
If we know the string literal that is being used as a format string then
we probably don't need to warn about it being const or not.

This is sort of a lazy approach.  We should still warn about other problems
with a known format string.  That's something TODO later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>