check_indenting: warn about missing curly braces and bad indenting
This check tests if each statement is indented the same as the statement
before it.  Also if the it's not the same but it is the same as the inside
of a previous if/else statement or loop then suggest adding curly braces.

This test adds over a thousand new warnings but they are very few false
positives.  Bad indenting does really hurt readability so I feel ok warning
about it.  Ideally checkpatch.pl would warn about these but you really
need to parse C at a deeper level to do it correctly so it's not possible
for checkpatch.pl.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>