extra: ignore more __in_fake_assign data
We're storing a lot of data which we don't use.  The bad thing is that we
do use a tiny bit of the equivalent data links, so this change *does*
introduce a few false positives.  The code looks like this:

	stuct foo *ops = some->really->long->name->ops;

	if (some->really->long->name->ops->foo)
		ops->foo();

All the false positives were messy like this so I am ok with the new false
positives.  People should write code consistently.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>