param_used: store when a parameter is used so we don't pass unused info
I'm preparing to commit a patch which records when data is set to zero
because it is allocated with kzalloc() much more accurately.  This means
that the database gets huge.  Most of the data is unused as well.

This patch allows us to tell which parameters are unused.  If a parameter
is unused and the value is zero then I filter it out.  I could maybe filter
out all unused data, but I'm not sure enough that this code is bug free to
do that.

The way this patch works is that if we try to look up the value of a
variable using get_state() then that variable is considered used.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>